requirements.txt
claude-agent-sdk>=0.1.46
anthropic>=0.28.0
aiogram==3.7.0
telethon>=1.35.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
click>=8.1.0
pyyaml>=6.0.1
aiohttp>=3.9.0
aiofiles>=23.2.0
aiosqlite>=0.20.0
python-dateutil>=2.9.0
pytz>=2024.1
pdfminer.six>=20221105
beautifulsoup4>=4.12.0
